+++ apps/wallet/src/main/scala/org/lfdecentralizedtrust/splice/wallet/UserWalletManager.scala
@@ -7,7 +7,11 @@ import org.apache.pekko.stream.Materializer
 import org.lfdecentralizedtrust.splice.codegen.java.splice.amulet as amuletCodegen
 import org.lfdecentralizedtrust.splice.codegen.java.splice.wallet.install.WalletAppInstall
 import org.lfdecentralizedtrust.splice.config.AutomationConfig
-import org.lfdecentralizedtrust.splice.environment.{SpliceLedgerClient, RetryProvider}
+import org.lfdecentralizedtrust.splice.environment.{
+  PackageVersionSupport,
+  RetryProvider,
+  SpliceLedgerClient,
+}
 import org.lfdecentralizedtrust.splice.environment.ledger.api.DedupDuration
 import org.lfdecentralizedtrust.splice.migration.DomainMigrationInfo
 import org.lfdecentralizedtrust.splice.scan.admin.api.client.BftScanConnection
@@ -52,6 +56,7 @@ class UserWalletManager(
     storage: Storage,
     retryProvider: RetryProvider,
     scanConnection: BftScanConnection,
+    packageVersionSupport: PackageVersionSupport,
     override val loggerFactory: NamedLoggerFactory,
     domainMigrationInfo: DomainMigrationInfo,
     participantId: ParticipantId,
@@ -227,6 +232,7 @@ class UserWalletManager(
       userRetryProvider,
       userLoggerFactory,
       scanConnection,
+      packageVersionSupport,
       domainMigrationInfo,
       participantId,
       ingestFromParticipantBegin,